third_party/WebKit/LayoutTests/media/video-seek-no-src-exception-expected.txt

Test that seeking video with no 'src' attribute throws an INVALID_STATE_ERR exception.

EXPECTED (video.networkState == '0') OK
EXPECTED (video.readyState == '0') OK
TEST(video.currentTime = 1.1) THROWS(DOMException.INVALID_STATE_ERR: Failed to set the 'currentTime' property on 'HTMLMediaElement': The element's readyState is HAVE_NOTHING.) OK
END OF TEST
